feat(courses): Add paid status to CourseCard and PaymentsModal
Introduces a `isPaid` property to courses and displays an 'Euro' icon on the CourseCard if the course is marked as paid.

The PaymentsModal now also displays the total amount of courses in the filtered list, in addition to the total of selected courses. This provides a more comprehensive view of financial data within the modal.